Refactored nodejs tests into their own file
- this way we don't have to test if the platform is nodejs all the
the time
